requirements.txt
python-dotenv
urllib3
pydantic
numpy
requests
lxml
python-dateutil
# jinja2
# fastapi[standard]
# uvicorn
# neo4j
biopython
nltk
sentence-transformers
# pandas
openpyxl
python-multipart
pydantic-ai
transformers
torch
llama-cpp-python